+++ src/app/animations-page/animations-page.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { SectionModule, ButtonModule } from '@swimlane/ngx-ui';
+
+import { PrismModule } from '../common/prism/prism.module';
+
+import { AnimationsPageComponent } from './animations-page.component';
+import { AnimationsPageRoutingModule } from './animations-page.routing.module';
+
+@NgModule({
+  declarations: [AnimationsPageComponent],
+  imports: [CommonModule, PrismModule, SectionModule, ButtonModule, AnimationsPageRoutingModule]
+})
+export class AnimationsPageModule {}
