+++ src/app/animations-page/animations-page.routing.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+
+import { AnimationsPageComponent } from './animations-page.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: AnimationsPageComponent
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class AnimationsPageRoutingModule {}
