+++ src/app/common/prism/prism.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { PrismComponent } from './prism.component';
+
+const declarations = [PrismComponent];
+
+@NgModule({
+  declarations,
+  exports: declarations,
+  imports: [CommonModule]
+})
+export class PrismModule {}
